backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
python-multipart
python-dotenv
